Refactor : socket 패키지에 interceptor 패키지 생성
- interceptor 패키지를 도입합니다.
- 건호님이 작성하신 SocketUtil 코드가 security보다 socket interceptor에 가까운 것 같아 패키지를 security에서 socket의 interceptor로 옮겼습니다.
