+++ src/main/java/nbbang/com/nbbang/global/socket/interceptor/SocketIdUtil.java
@@ -1,4 +1,4 @@
-package nbbang.com.nbbang.global.security.context;
+package nbbang.com.nbbang.global.socket.interceptor;
 
 import org.springframework.http.server.ServerHttpRequest;
 import org.springframework.messaging.Message;
